+++ bricks/_common/micropython.c
@@ -307,6 +307,7 @@ pbio_error_t pbsys_main_program_validate(pbsys_main_program_t *program) {
             #endif
             #if PBSYS_CONFIG_FEATURE_BUILTIN_USER_PROGRAM_EV3_APPS
             case PBIO_PYBRICKS_USER_PROGRAM_ID_EV3_MOTOR_BUTTON_CONTROL:
+            case PBIO_PYBRICKS_USER_PROGRAM_ID_EV3_MOTOR_IR_CONTROL:
                 return PBIO_SUCCESS;
             #endif
             default:
@@ -403,6 +404,10 @@ void pbsys_main_run_program(pbsys_main_program_t *program) {
         case PBIO_PYBRICKS_USER_PROGRAM_ID_EV3_MOTOR_BUTTON_CONTROL:
             pb_package_pybricks_init(false);
             pyexec_frozen_module("_ev3_motor_button_control.py", false);
+            break;
+        case PBIO_PYBRICKS_USER_PROGRAM_ID_EV3_MOTOR_IR_CONTROL:
+            pb_package_pybricks_init(false);
+            pyexec_frozen_module("_ev3_motor_ir_control.py", false);
             break;
         #endif