+++ lib/pbio/sys/hmi_ev3_ui.c
@@ -85,6 +85,7 @@ static pbsys_hmi_ev3_ui_t state;
  */
 static const char *apps[] = {
     " Motor Button Control",
+    " Motor IR Control",
 };
 
 /**
@@ -282,6 +283,9 @@ pbsys_hmi_ev3_ui_action_t pbsys_hmi_ev3_ui_handle_button(pbio_button_flags_t but
             switch (state.selection[PBSYS_HMI_EV3_UI_TAB_APPS]) {
                 case 0:
                     *payload = PBIO_PYBRICKS_USER_PROGRAM_ID_EV3_MOTOR_BUTTON_CONTROL;
+                    return PBSYS_HMI_EV3_UI_ACTION_RUN_PROGRAM;
+                case 1:
+                    *payload = PBIO_PYBRICKS_USER_PROGRAM_ID_EV3_MOTOR_IR_CONTROL;
                     return PBSYS_HMI_EV3_UI_ACTION_RUN_PROGRAM;
                 default:
                     // Other apps not yet implemented.