+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `pybricks.pupdevices.MarioHub` to control it as a peripheral. It
   cannot be used as a standalone device since it cannot ne updated.
 - Added `DriveBase.hold()` method ([support#2621]).
+- Added `Motor Button Control` and `Motor IR Control` on EV3 apps tab.
 
 ### Changed
 - The EV3 Color Sensor now returns `Color.NONE` instead of `None` when no color